custome hook introduced, the fetch features goes to independent file
