+++ github-users/src/components/GitUserList.js
@@ -2,44 +2,14 @@ import React, { useEffect, useState } from 'react'
 import User from './User'
 import classes from './GitUsers.module.css'
 import spinner from '../assets/spinner1.jpg'
+import useFetch from './hooks/useFetch'
 
+
+const url = 'https://api.github.com/users'
 
 const GitUserList = () => {
-    const [users, setUsers] = useState([]);
-    const [loading, setLoading] = useState(false);
-    const [error, setError] = useState(false);
-
-
-    const fetchGitUsers= async()=>{
-        try {
-
-            setLoading(true);
-            setError(false);
-                const response = await fetch('https://api.github.com/users')
-                console.log(response)
-                if(!response.ok){
-                    throw new Error('Something went wrong')
-
-                }
-                const data = await response.json();
-                setUsers(data)
-                setLoading(false);
-    
-        } catch (error) {
-            console.log(error.message)
-            setError(true)
-            setLoading(false);
-            
-        }
-
-    }
-
-    useEffect(() => {
-        // fetch('https://api.github.com/users')
-        // .then(res => res.json())
-        // .then(data => setUsers(data))
-        fetchGitUsers();
-    }, [])
+  const {data,error, loading} = useFetch(url)
+   
 
   return (
     <div className="container   --bg-primary">
@@ -49,7 +19,7 @@ const GitUserList = () => {
     <div className="--grid-25 --py">
 
     {error ? (<p className="--text-center --text-light">Something went wrong</p>) : (
-        users.map(user => (
+        data.map(user => (
         <User key={user.id} user={user}/>
 
     ))
